+++ src/types.js
@@ -1000,12 +1000,17 @@ class ResourceNode {
   getTypeInfo() {
     const namespace = TypeInfo.FHIR;
 
-    // TODO: Here we should use property index which we will extract from the specification
-
-    if (this.path.indexOf('.') === -1) {
+    if (/^System\.(.*)$/.test(this.path)) {
+      return new TypeInfo({namespace: TypeInfo.System, name: RegExp.$1});
+    } else if (this.path.indexOf('.') === -1) {
       return new TypeInfo({namespace, name: this.path});
     }
-    return TypeInfo.createByValueInNamespace({namespace, value: this.data});
+
+    if (!TypeInfo.model) {
+      return TypeInfo.createByValueInNamespace({namespace, value: this.data});
+    }
+
+    return new TypeInfo({namespace, name: 'BackboneElement'});
   }
 
   toJSON() {
@@ -1060,6 +1065,9 @@ class TypeInfo {
     this.namespace = namespace;
   }
 
+  // The "model" data object specific to a domain, e.g. R4.
+  static model = null;
+
   /**
    * Checks for equality with another TypeInfo object, or that another TypeInfo
    * object specifies a superclass for the type specified by this object.
@@ -1067,9 +1075,20 @@ class TypeInfo {
    * @return {boolean}
    */
   is(other) {
-    // TODO: Here we should use type hierarchy index which we will extract from the specification
-    return other instanceof TypeInfo && this.name === other.name
-      && (!this.namespace || !other.namespace || this.namespace === other.namespace);
+    if (other instanceof TypeInfo
+      && (!this.namespace || !other.namespace || this.namespace === other.namespace)) {
+      if (TypeInfo.model && (!this.namespace || this.namespace === TypeInfo.FHIR)) {
+        let name = this.name;
+        do {
+          if (name === other.name) {
+            return true;
+          }
+        } while ((name = TypeInfo.model.type2Parent[name]));
+      } else {
+        return this.name === other.name;
+      }
+    }
+    return false;
   }
 }
 
@@ -1142,10 +1161,29 @@ function isFn(coll, typeInfo) {
   }
 
   if(coll.length > 1) {
-    throw new Error("Expected singleton on left side of is, got " + JSON.stringify(coll));
+    throw new Error("Expected singleton on left side of 'is', got " + JSON.stringify(coll));
   }
 
   return TypeInfo.fromValue(coll[0]).is(typeInfo);
+}
+
+/**
+ * Implementation of function "as(type : type specifier)" and operator "as"
+ * (see http://hl7.org/fhirpath/#as-type-specifier)
+ * @param {Array<*>} coll - input collection
+ * @param {TypeInfo} typeInfo
+ * @return {Array<*>}
+ */
+function asFn(coll, typeInfo) {
+  if(coll.length === 0) {
+    return [];
+  }
+
+  if(coll.length > 1) {
+    throw new Error("Expected singleton on left side of 'as', got " + JSON.stringify(coll));
+  }
+
+  return TypeInfo.fromValue(coll[0]).is(typeInfo) ? coll : [];
 }
 
 module.exports = {
@@ -1159,5 +1197,6 @@ module.exports = {
   ResourceNode: ResourceNode,
   TypeInfo: TypeInfo,
   typeFn,
-  isFn
+  isFn,
+  asFn
 };